requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
celery==5.3.6
redis==5.0.1
openai==1.10.0
langchain==0.1.0
chromadb==0.4.22
twilio==8.11.0
slack-sdk==3.26.0
google-cloud-speech==2.23.0
gunicorn==21.2.0
